Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@3214 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ testjava/src/TestBeans.java
@@ -70,7 +70,7 @@ public class TestBeans
 
 
 
-        showInt(some.getObjInteger());
+        showInt(some.getObjInteger().intValue());
     }
 
     public static void showInt(int i)
